+++ scripts/test_bongocat_8.sh
@@ -15,6 +15,7 @@ cp $OG_CONFIG $CONFIG
 
 sed -i -E 's/^cat_height=[0-9]+/cat_height=96/' "$CONFIG"
 sed -i -E 's/^overlay_height=[0-9]+/overlay_height=128/' "$CONFIG"
+sed -i -E 's/^enable_antialiasing=[0-9]+/enable_antialiasing=1/' "$CONFIG"
 
 if [[ $# -ge 1 ]]; then
     PID="$1"
@@ -24,7 +25,7 @@ if [[ $# -ge 1 ]]; then
     echo "[TEST] Using provided PID = $PID"
 else
     echo "[TEST] Starting program..."
-    "$PROGRAM" --config "$CONFIG" --ignore-running --strict &
+    "$PROGRAM" --config "$CONFIG" --ignore-running &
     PID=$!
     echo "[TEST] Program PID = $PID"
     sleep 5
@@ -57,6 +58,7 @@ sleep 7
 
 echo "[TEST] Change overlay settings"
 echo "[INFO] Set overlay_height"
+sed -i -E 's/^cat_height=[0-9]+/cat_height=96/' "$CONFIG"
 sed -i -E 's/^overlay_height=[0-9]+/overlay_height=100/' "$CONFIG"
 echo "[INFO] Send SIGUSR2"
 kill -USR2 "$PID" # Reload config
@@ -87,6 +89,19 @@ echo "[INFO] Send SIGUSR2"
 kill -USR2 "$PID" # Reload config
 sleep 10
 
+echo "[INFO] Set max size"
+sed -i -E 's/^cat_height=[0-9]+/cat_height=1024/' "$CONFIG"
+sed -i -E 's/^overlay_height=[0-9]+/overlay_height=2560/' "$CONFIG"
+echo "[INFO] Send SIGUSR2"
+kill -USR2 "$PID" # Reload config
+sleep 15
+echo "[INFO] Set min size"
+sed -i -E 's/^cat_height=[0-9]+/cat_height=16/' "$CONFIG"
+sed -i -E 's/^overlay_height=[0-9]+/overlay_height=32/' "$CONFIG"
+echo "[INFO] Send SIGUSR2"
+kill -USR2 "$PID" # Reload config
+sleep 15
+
 sleep 20
 
 # --- verify running ---
@@ -97,6 +112,9 @@ else
     exit 1
 fi
 
+echo "[INFO] Set overlay_height"
+sed -i -E 's/^cat_height=[0-9]+/cat_height=128/' "$CONFIG"
+sed -i -E 's/^overlay_height=[0-9]+/overlay_height=256/' "$CONFIG"
 echo "[TEST] Set Monitor"
 echo "[INFO] Set monitor"
 sed -i -E 's/^monitor=.*/monitor=HDMI-A-1/' "$CONFIG"